feat(ml): integrate advanced alpha features and automated retraining
This commit adds 4 new volatility/pullback features (Synthetic Order Flow, Parkinson Volatility, Gap vs Grind, Dynamic Micro-Pullbacks) to the GodLevelSignalGenerator. It also correctly parses sliding OHLCV dataframes into the legacy extract_features logic, and adds a dedicated train_god_level_models.py background worker directly hooked into the apex_ctl.sh hot-reloading manager.

+++ risk/pretrade_risk_gateway.py
@@ -66,7 +66,9 @@ class PreTradeRiskGateway:
         if not self.config.enabled:
             return PreTradeDecision(True, "disabled", "Pre-trade gateway disabled")
 
-        qty = int(abs(quantity))
+        # Use float to support fractional crypto shares (e.g. 0.1218 BTC).
+        # int() would truncate fractional positions to 0 → incorrectly blocked.
+        qty = float(abs(quantity))
         px = float(price)
         eval_price = px if px > 0 else float(reference_price or 0.0)
